pyproject.toml
[project]
name = "mlops-project"
version = "0.1.0"
description = ""
authors = [
    {name = "ruanwko",email = "artem.borisov@raiffeisen.ru"}
]
readme = "README.md"
requires-python = "==3.12.4"
dependencies = [
    "lightning (>=2.6.0,<3.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "mlflow (>=3.8.1,<4.0.0)",
    "dvc (>=3.65.0,<4.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "dvc-gdrive (>=3.0.1,<4.0.0)",
    "pyopenssl (>=25.3.0,<26.0.0)",
    "gdown (>=5.2.0,<6.0.0)",
    "flagembedding (>=1.3.5,<2.0.0)",
    "dill (>=0.4.0,<0.5.0)",
    "graphql-core (>=3.2.7,<4.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "fire (>=0.7.1,<0.8.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "isort (>=7.0.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
